+++ src/main/java/com/feiyu/discord/sg/tavern/commands/InviteLinkCommand.java
@@ -25,10 +25,11 @@ public class InviteLinkCommand extends ListenerAdapter {
     @Override
     public void onSlashCommandInteraction(@NotNull SlashCommandInteractionEvent event) {
         if (event.getName().equals("invite")) {
+            log.info("Command - invite - {}", event.getUser().getId());
+            
             Member member = event.getMember();
             Guild guild = event.getGuild();
             TextChannel landingChannel = guild.getTextChannelById(valuesConfig.getRulesChannelId());
-            log.info("InviteLinkCommand.invitelink by : {}", member.getUser().getName());
             
             Invite invite = landingChannel.createInvite()
                     .setMaxAge(2L, TimeUnit.DAYS)