+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.2] - 2026-02-26
+- Bug fix on sendGpt component
+- deferred reply for `/extractevent`
+
+## [2.2.1] - 2026-02-25
+- Bug fix on `/eventstatus` and sendGpt component
+- Rollback on event scheduler's gpt component
+
 ## [2.2.0] - 2026-02-01
 - Major version change for JDA to 6.3.0 from 5.6.1
 - Bug fix on `/eventstatus` and sendGpt component